+++ crates/bitang/src/engine/draw.rs
@@ -1,9 +1,9 @@
-use crate::control::controls::{Control, Globals};
+use crate::control::controls::{Control};
 use crate::engine::camera::Camera;
 use crate::engine::pass::Pass;
 use crate::engine::render_object::RenderObject;
 use crate::engine::scene::Scene;
-use super::{FrameContext, RenderPassContext, Viewport};
+use super::{FrameContext, Globals, RenderPassContext, Viewport};
 use anyhow::{ensure, Result};
 use glam::{Mat4, Vec2, Vec3};
 use std::rc::Rc;
