+++ crates/bitang/src/engine/mod.rs
@@ -44,3 +44,5 @@ pub use core::shader::GlobalUniformMapping;
 pub use core::SIMULATION_STEP_SECONDS;
 
 pub use core::context::{GpuContext, Viewport, FrameContext, RenderPassContext, ComputePassContext};
+
+pub use core::globals::{Globals, GlobalType};
